Rename `BaseModule` to `Module`
