+++ src/main/java/decodex/modules/hex/HexEncoder.java
@@ -1,12 +1,12 @@
 package decodex.modules.hex;
 
 import decodex.data.Data;
-import decodex.modules.BaseModule;
+import decodex.modules.Module;
 
 /**
  * The HexEncoder class converts the input string to hexadecimal bytes.
  */
-public class HexEncoder extends BaseModule {
+public class HexEncoder extends Module {
 
     public static final String MODULE_NAME = "hexencode";
     public static final String MODULE_DESCRIPTION = "Converts the input string to hexadecimal bytes";
